Trade Project
This project has code to send thousands of trades that are being sent to a single store, through any method
of transmission. The objective is to develop a trade store that organises and stores each trade in a
specific sequence with below validations to be implemented:
1. If a trade with a lower version is received during transmission, the store will reject it and
generate an exception. Trades with the same version will replace the current record.
2. The store must reject any trade that has a maturity date earlier than today's date.
3. When a trade's maturity date is surpassed, the store should automatically mark the trade as
expired.